+++ certbot/CHANGELOG.md
@@ -15,7 +15,8 @@ Certbot adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
-*
+* Removed a CryptographyDeprecationWarning that was being displayed to users
+  when checking OCSP status.
 
 More details about these changes can be found on our GitHub repo.